\name{imgMat}
\alias{imgMat}

\title{
Convert a PNG file to a logical matrix
}
\description{
Convert a PNG file to a logical matrix
}
\usage{
imgMat(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Image path, ended with .png 
}
}

\value{
A logical matrix indicating white or black pixels of the image
}